Rename misspelled files
