+++ linkedlist/linkedlist_endpoint_helpers.py
