+++ linkedlist/linkedlist_main.py
@@ -5,7 +5,7 @@ import uvicorn
 from fastapi import FastAPI
 from fastapi.responses import RedirectResponse
 
-import linkedlist_enpoint_helpers as linkedlist_help
+import linkedlist_endpoint_helpers as linkedlist_help
 import linkedlist_request_classes as linkedlist_req
 from common import config_info
